# controllers/chips.js
const db = require('../database/mongodb');
const ObjectId = require('mongodb').ObjectId;

const getData = async (req, res, next) => {
  const result = await db.getDb().db().collection('chips').find();
  result.toArray().then((lists) => {
    res.setHeader('Content-Type', 'application/json');
    res.status(200).json(lists); 
  });
};
const getDataById =  async (req, res, next) => {
  try {
    const id =  new ObjectId(req.params.id);
    const result = await db.getDb().db().collection('chips').findOne({ _id: id });
    res.setHeader('Content-Type', 'application/json');
    if (result) {
      res.status(200).json(result);
    } else {
      res.status(404).json({ message: 'Professional not found' });
    }
  }
  catch (error) {
    next(error);
  }
}
//create function for  router.post(professionalController.createData);
const createData = async (req, res, next) => {
  try {
    const data = {
      brand: req.body.brand,
      flavor: req.body.flavor,
      type: req.body.type,
      production_cost: req.body.cost,
      retail_price: req.body.price,
      nutritional_value: req.body.nutritional_value,
      ingredients: req.body.ingredients,
      image: req.body.image 
    }
    const result = await db.getDb().db().collection('chips').insertOne(data);
    if(result.acknowledged){
      res.status(200).json({ message: 'Professional inserted successfully' });
    } else {
      res.status(404).json({ message: 'Professional not inserted' });
    }
  }
  catch (error) {
    next(error);
  }
}
// create function for router.put('/:id', professionalController.updateData);
const updateData = async (req, res, next) => {
  try {
    const id = new ObjectId(req.params.id);
    const data = {
      "brand": req.body.brand,
      "flavor": req.body.flavor,
      "type": req.body.type,
      "production_cost": req.body.cost,
      "retail_price": req.body.price,
      "nutritional_value": req.body.nutritional_value,
      "ingredients": req.body.ingredients,
      "image": req.body.image 
    }
    const result = await db.getDb().db().collection('chips').replaceOne(
      { _id: id },
      data
    );
    if(result.modifiedCount > 0){
      res.status(200).json({ message: 'Professional updated successfully' });
    } else {
      res.status(404).json({ message: 'Professional not found' });
    }
  }
  catch (error) {
    next(error);
  }
}

//create function for  router.delete('/:id', professionalController.deleteData);
const deleteData = async (req, res, next) => {
  try {
    const id = new ObjectId(req.params.id);
    const result = await db.getDb().db().collection('chips').deleteOne({ _id: id });
  if(result.deletedCount > 0){
      res.status(200).json({ message: 'Professional deleted successfully' });
    } else {
      res.status(404).json({ message: 'Professional not found' });
    }
  }
  catch(error){
    next(error);
  }
}
module.exports = { getData,  getDataById, updateData, createData, deleteData };